requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
tenacity>=8.2.3